Class type PPrintEngine.output

method char : char -> unit

char c sends the character c to the output channel.

method substring : string -> int -> int -> unit

substring s ofs len sends the substring of s delimited by the offset ofs and the length len to the output channel.